Major update of discord.js package
Also added additional packages that are no longer included in discord.js
Made som inor changes to the actual app code so it at least starts and
commands being processed.
The big thing left to fix is the preach and insult speech commands which
has to be rewritten with the new api in mind.